# docs/swagger.yaml
definitions:
  domain.Filter:
    properties:
      end_date:
        type: string
      endDate:
        type: string
      limit:
        type: integer
      offset:
        type: integer
      price:
        type: integer
      service_name:
        type: string
      start_date:
        type: string
      startDate:
        type: string
      user_id:
        type: string
    type: object
  domain.Subscription:
    properties:
      end_date:
        type: string
      price:
        type: integer
      service_name:
        type: string
      start_date:
        type: string
      user_id:
        type: string
    type: object
  domain.SubscriptionInput:
    properties:
      end_date:
        type: string
      price:
        type: integer
      service_name:
        type: string
      start_date:
        type: string
      user_id:
        type: string
    type: object
info:
  contact: {}
paths:
  /api/subscriptions:
    delete:
      consumes:
      - application/json
      description: Удаление подписки по user_id и service_name
      parameters:
      - description: ID пользователя
        in: query
        name: user_id
        required: true
        type: string
      - description: Название сервиса
        in: query
        name: service_name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: deleted
          schema:
            type: string
        "400":
          description: bad request
          schema:
            type: string
        "404":
          description: not found
          schema:
            type: string
        "500":
          description: internal error
          schema:
            type: string
      summary: Удалить подписку
      tags:
      - subscriptions
    get:
      consumes:
      - application/json
      description: Поиск подписок по фильтру
      parameters:
      - description: ID пользователя
        in: query
        name: user_id
        type: string
      - description: Название сервиса
        in: query
        name: service_name
        type: string
      - description: Цена
        in: query
        name: price
        type: integer
      - description: Дата начала MM-YYYY
        in: query
        name: start_date
        type: string
      - description: Дата окончания MM-YYYY
        in: query
        name: end_date
        type: string
      - description: Лимит
        in: query
        name: limit
        type: integer
      - description: Смещение
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Subscription'
            type: array
        "400":
          description: bad request
          schema:
            type: string
        "500":
          description: internal error
          schema:
            type: string
      summary: Получить список подписок
      tags:
      - subscriptions
    post:
      consumes:
      - application/json
      description: Добавить новую подписку
      parameters:
      - description: Данные подписки
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/domain.SubscriptionInput'
      produces:
      - application/json
      responses:
        "201":
          description: created
          schema:
            type: string
        "400":
          description: bad request
          schema:
            type: string
        "500":
          description: internal error
          schema:
            type: string
      summary: Создать подписку
      tags:
      - subscriptions
    put:
      consumes:
      - application/json
      description: Обновление данных подписки по user_id и service_name
      parameters:
      - description: Данные подписки
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/domain.SubscriptionInput'
      produces:
      - application/json
      responses:
        "200":
          description: updated
          schema:
            type: string
        "400":
          description: bad request
          schema:
            type: string
        "500":
          description: internal error
          schema:
            type: string
      summary: Обновить подписку
      tags:
      - subscriptions
  /api/subscriptions/summary:
    post:
      consumes:
      - application/json
      description: Сводная информация по подпискам за период
      parameters:
      - description: Фильтр с датами
        in: body
        name: filter
        required: true
        schema:
          $ref: '#/definitions/domain.Filter'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: integer
            type: object
        "400":
          description: bad request
          schema:
            type: string
        "500":
          description: internal error
          schema:
            type: string
      summary: Получить сумму подписок за период
      tags:
      - subscriptions
swagger: "2.0"